move class Product to items.py
